refactor: split instantiators into modules, rename to TcpInstantiator
Move the instantiator implementations into src/instantiators/* (base,
command, tcp, test) and rename HTTPInstantiator to TcpInstantiator to
match the TCP-on-grpcPort transport. Revert the earlier shutdown change.

Squashed from 56ab448.
